requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
PyJWT==2.1.0
pymongo==3.12.0
mysql-connector-python==8.0.26
redis==4.1.0
python-dotenv==0.19.0
qrcode==7.3.1
opencv-python==4.5.5.62
pandas==1.3.5
matplotlib==3.5.0
bcrypt==4.0.1
gunicorn==20.1.0
pymysql==1.0.3
Werkzeug==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
